refactor DataSource
* make DataSource a spring bean
* move checkJavaAndMySQLOffsets() into GenericRepository
* necessary consequential changes
